pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipescope"
version = "0.1.0"
description = "CI/CD security enumeration tool for DevOps pipelines"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.12.0",
  "pyyaml>=6.0.1",
  "rich>=13.7.1",
  "detect-secrets>=1.5.0",
  "requests>=2.28.0",
  "reportlab>=4.0.0",
]

[project.scripts]
pipescope = "pipescope.cli:main"

[tool.setuptools.packages.find]
include = ["pipescope*"]
exclude = ["repos*"]

